requirements.txt
torch>=2.2
monai>=1.4
nibabel>=5.2
numpy>=1.26
scikit-image>=0.24
scikit-learn>=1.5
tqdm>=4.66
matplotlib>=3.8
tensorboard>=2.16
pydantic>=2.8
omegaconf>=2.3
rich>=13.7
kaggle>=1.6.14
pyyaml>=6.0.2
gradio>=4.44
pytest>=8.2
pytest-cov>=5.0
mypy>=1.11
ruff>=0.6.9
black>=24.8
isort>=5.13
